streamline camera library workflow
This removes Windows UWP support effectively and merges the Win32 and Linux implementations back to the main project.
