--- Vignette.Application.Camera.Linux/Vignette.Application.Camera.Linux.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>Vignette.Application.Camera.Linux</PackageId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Vignette.Application.Camera\Vignette.Application.Camera.csproj" />
-  </ItemGroup>
-
-</Project>
