+++ Vignette.Application.Camera.Tests/CameraManagerTest.cs
@@ -21,7 +21,7 @@ namespace Vignette.Application.Camera.Tests
         [SetUp]
         public void SetUp()
         {
-            manager = CreateSuitableManager(scheduler = new Scheduler());
+            manager = CameraManager.CreateSuitableManager(scheduler = new Scheduler());
         }
 
         [Test]
@@ -33,21 +33,6 @@ namespace Vignette.Application.Camera.Tests
 
             // This requires a physical device to be connected to succeed.
             Assert.IsTrue(manager.CameraDeviceNames.Any());
-        }
-
-        protected virtual CameraManager CreateSuitableManager(Scheduler scheduler)
-        {
-            switch (RuntimeInfo.OS)
-            {
-                case RuntimeInfo.Platform.Windows:
-                    return new LegacyWindowsCameraManager(scheduler);
-
-                case RuntimeInfo.Platform.Linux:
-                    return new LinuxCameraManager(scheduler);
-
-                default:
-                    throw new PlatformNotSupportedException();
-            }
         }
     }
 }
