--- Vignette.Application.Camera.Windows.Legacy/Vignette.Application.Camera.Windows.Legacy.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>Vignette.Application.Camera.Windows.Legacy</PackageId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="DirectShowLib.Standard" Version="2.1.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Vignette.Application.Camera\Vignette.Application.Camera.csproj" />
-  </ItemGroup>
-
-</Project>
