--- Vignette.Application.Camera.Windows.Tests/Vignette.Application.Camera.Windows.Tests.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows10.0.17763.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="NUnit" Version="3.13.0" />
-    <PackageReference Include="NUnit3TestAdapter" Version="3.17.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Vignette.Application.Camera.Tests\Vignette.Application.Camera.Tests.csproj" />
-    <ProjectReference Include="..\Vignette.Application.Camera.Windows\Vignette.Application.Camera.Windows.csproj" />
-  </ItemGroup>
-
-</Project>
