--- Vignette.Application.Camera.Windows/Vignette.Application.Camera.Windows.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0-windows10.0.17763.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <PackageId>Vignette.Application.Camera.Windows</PackageId>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Vignette.Application.Camera\Vignette.Application.Camera.csproj" />
-  </ItemGroup>
-
-</Project>
