+++ Vignette.Application.Camera/Platform/LinuxCameraManager.cs
